+++ src/server.cpp
@@ -5,8 +5,8 @@
 using namespace std;
 
 float BLOCK_SIZE = 1;
-uint BLOCK_COUNT = 2;
-uint PRINT_COUNT = 20;
+size_t BLOCK_COUNT = 2;
+size_t PRINT_COUNT = 20;
 Logger logger;
 vector<string> tokenizedQuery;
 ParsedQuery parsedQuery;
